More fixes and tweeks

+++ hack/ose_images_rebuild.sh
@@ -18,7 +18,8 @@
 #   git
 
 BASE_GIT_REPO="git@github.com:openshift/ose.git"
-DIST_GIT_BRANCH="rhaos-3.1-rhel-7-candidate"
+DIST_GIT_BRANCH="rhaos-3.1-rhel-7"
+#DIST_GIT_BRANCH="rhaos-3.1-rhel-7-candidate"
 # format:
 # dist-git_name	image_dependency dist-git_branch git_repo git_path
 base_images_list="
@@ -51,21 +52,39 @@ openshift-sti-php-docker openshift-sti-base-docker ${DIST_GIT_BRANCH} https://gi
 openshift-sti-python-docker openshift-sti-base-docker ${DIST_GIT_BRANCH} https://github.com/openshift/sti-python sti-python/3.3
 openshift-sti-ruby-docker openshift-sti-base-docker ${DIST_GIT_BRANCH} https://github.com/openshift/sti-ruby sti-ruby/2.0
 "
+# format:
+# dist-git_name	image_dependency dist-git_branch git_repo git_path
+logging_images_list="
+logging-auth-proxy-docker None ${DIST_GIT_BRANCH} https://github.com/openshift/sti-base sti-base
+logging-deployment-docker None ${DIST_GIT_BRANCH} https://github.com/openshift/mongodb mongodb/2.4
+logging-elasticsearch-docker None ${DIST_GIT_BRANCH} https://github.com/openshift/mysql mysql/5.5
+logging-fluentd-docker None ${DIST_GIT_BRANCH} https://github.com/openshift/postgresql postgresql/9.2
+logging-kibana-docker openshift-sti-base-docker ${DIST_GIT_BRANCH} https://github.com/openshift/sti-nodejs sti-nodejs/0.10
+"
+# format:
+# dist-git_name	image_dependency dist-git_branch git_repo git_path
+metrics_images_list="
+metrics-cassandra-docker None ${DIST_GIT_BRANCH} https://github.com/openshift/origin-metrics.git origin-metrics/cassandra
+metrics-deployer-docker None ${DIST_GIT_BRANCH} https://github.com/openshift/origin-metrics.git origin-metrics/deployer
+metrics-hawkular-metrics-docker None ${DIST_GIT_BRANCH} https://github.com/openshift/origin-metrics.git origin-metrics/hawkular-metrics
+metrics-heapster-docker None ${DIST_GIT_BRANCH} https://github.com/openshift/origin-metrics.git origin-metrics/heapster
+"
 
 usage() {
-  echo "Usage `basename $0` <action> <option>" >&2
+  echo "Usage `basename $0` [action] [--group <group>] <option>" >&2
   echo >&2
   echo "Actions:" >&2
-  echo "  git_update_base  - Clone git and dist-git, bump release, compare (non-s2i images)" >&2
-  echo "  git_update_s2i   - Clone git and dist-git, bump release, compare (s2i images)" >&2
-  echo "  build_container_base - Clone dist-git, build containers (non-s2i images)" >&2
-  echo "  build_container_s2i  - Clone dist-git, build containers (s2i images)" >&2
-  echo "  everything_base  - git_update,  build_container (non s2i images)" >&2
-  echo "  everything_s2i   - git_update,  build_container (s2i images)" >&2
+  echo "  git_update  - Clone git and dist-git, bump release, compare" >&2
+  echo "  build_container - Clone dist-git, build containers" >&2
+  echo "  everything  - git_update,  build_container" >&2
+  echo >&2
+  echo "Groups:" >&2
+  echo "  base s2i metrics logging" >&2
   echo >&2
   echo "Options:" >&2
   echo "  -h, --help          :: Show this options menu" >&2
   echo "  -v, --verbose       :: Be verbose" >&2
+  echo "  -f, --force         :: Force: git_update always select continue" >&2
   echo "  --version [version] :: Change Dockerfile version e.g. 3.1.1.2" >&2
   echo "  --release [version] :: Change Dockerfile release e.g. 3" >&2
   echo "  --rhel [version]    :: Change Dockerfile RHEL version e.g. rhel7.2:7.2-35 or rhel7:latest" >&2
@@ -117,17 +136,16 @@ check_builds() {
     if grep -q -e "buildContainer (noarch) failed" -e "server startup error" ${line} ; then
       package=`echo ${line} | cut -d'.' -f1`
       echo "=== ${package} IMAGE BUILD FAILED ==="
-      echo " exiting"
       mv ${line} ${package}.watchlog done/
-      #break 9
+      echo "::${package}::" >> ${workingdir}/logs/finished
+      echo "::${package}::" >> ${workingdir}/logs/buildfailed
     fi
     if grep -q -e "buildContainer (noarch) completed successfully" ${line} ; then
       package=`echo ${line} | cut -d'.' -f1`
       echo "==== ${package} IMAGE COMPLETED ===="
       if grep "No package" ${package}.watchlog ; then
         echo "===== ${package}: ERRORS IN COMPLETED IMAGE see above ====="
-        #echo "  exiting so we do not build more images based on this bad image"
-        #break 9
+        echo "::${package}::" >> ${workingdir}/logs/buildfailed
       fi
       echo "::${package}::" >> ${workingdir}/logs/finished
       mv ${line} ${package}.watchlog done/
@@ -164,29 +182,44 @@ check_build_dependencies() {
 build_image() {
   pushd "${workingdir}/${container}" >/dev/null
   check_build_dependencies
-  rhpkg container-build --repo http://file.rdu.redhat.com/sdodson/aos-unsigned.repo >> ${workingdir}/logs/${container}.buildlog 2>&1 &
-  #rhpkg container-build --scratch --repo http://file.rdu.redhat.com/sdodson/aos-unsigned.repo >> ${workingdir}/logs/${container}.buildlog 2>&1 &
-  echo -n "  Waiting for createContainer taskid ."
-  taskid=`grep createContainer ${workingdir}/logs/${container}.buildlog | awk '{print $1}' | sort -u`
-  while [ "${taskid}" == "" ]
-  do
-    echo -n "."
-    sleep 5
-    if grep -q -e "buildContainer (noarch) failed" -e "server startup error" ${workingdir}/logs/${container}.buildlog ; then
-      echo " error"
-      echo "=== ${container} IMAGE BUILD FAILED ==="
-      echo "  exiting"
-      #break 9
-    fi
+  failedcheck=`grep ${dependency} ${workingdir}/logs/buildfailed`
+  if [ "${failedcheck}" == "" ] ; then
+    rhpkg container-build --repo http://file.rdu.redhat.com/tdawson/repo/aos-unsigned-errata.repo >> ${workingdir}/logs/${container}.buildlog 2>&1 &
+    #rhpkg container-build --repo http://file.rdu.redhat.com/sdodson/aos-unsigned.repo >> ${workingdir}/logs/${container}.buildlog 2>&1 &
+    #rhpkg container-build --repo http://file.rdu.redhat.com/tdawson/repo/aos-unsigned-enabled-errata.repo >> ${workingdir}/logs/${container}.buildlog 2>&1 &
+    #rhpkg container-build --scratch --repo http://file.rdu.redhat.com/sdodson/aos-unsigned.repo >> ${workingdir}/logs/${container}.buildlog 2>&1 &
+    echo -n "  Waiting for createContainer taskid ."
     taskid=`grep createContainer ${workingdir}/logs/${container}.buildlog | awk '{print $1}' | sort -u`
-  done
-  echo " "
-  brew watch-logs ${taskid} >> ${workingdir}/logs/${container}.watchlog 2>&1 &
+    while [ "${taskid}" == "" ]
+    do
+      echo -n "."
+      sleep 5
+      taskid=`grep createContainer ${workingdir}/logs/${container}.buildlog | awk '{print $1}' | sort -u`
+      if grep -q -e "buildContainer (noarch) failed" -e "server startup error" ${workingdir}/logs/${container}.buildlog ; then
+        echo " error"
+        echo "=== ${container} IMAGE BUILD FAILED ==="
+        mv ${workingdir}/logs/${container}.buildlog done/
+        echo "::${package}::" >> ${workingdir}/logs/finished
+        echo "::${package}::" >> ${workingdir}/logs/buildfailed
+        taskid="FAILED"
+      fi
+    done
+    echo " "
+    if ! [ "${taskid}" == "FAILED" ] ; then
+      brew watch-logs ${taskid} >> ${workingdir}/logs/${container}.watchlog 2>&1 &
+    fi
+  else
+    echo "  depedency error: ${dependency} failed to build"
+    echo "=== ${container} IMAGE BUILD FAILED ==="
+    echo "::${package}::" >> ${workingdir}/logs/finished
+    echo "::${package}::" >> ${workingdir}/logs/buildfailed
+  fi
   popd >/dev/null
 }
 
 show_git_diffs() {
   pushd "${workingdir}/${container}" >/dev/null
+  echo "  ---- Checking Dockerfile changes ----"
   find . -name ".osbs*" -prune -o -name "Dockerfile*" -type f -print | while read line
   do
     diff --label ${line} --label ${path}/${line} -u ${line} ${workingdir}/${path}/${line} >> .osbs-logs/${line}.diff.new
@@ -196,21 +229,28 @@ show_git_diffs() {
     else
       echo "${newdiff}"
       echo " "
-      echo "(c)ontinue [replace old diff], (i)gnore [leave old diff], (q)uit [exit script] : "
-      read choice < /dev/tty
-      case ${choice} in
-        c | C | continue ) mv -f .osbs-logs/${line}.diff.new .osbs-logs/${line}.diff ; git add .osbs-logs/${line}.diff ; rhpkg commit -p -m "Updating dockerfile diff" ;;
-        i | I | ignore ) rm -f .osbs-logs/${line}.diff.new ;;
-        q | Q | quit ) break 5 ;;
-        * ) echo "${choice} not and option.  Assuming ignore" ; rm -f .osbs-logs/${line}.diff.new ;;
-        #* ) echo "${choice} not and option.  Assuming continue" ;  mv -f .osbs-logs/${line}.diff.new .osbs-logs/${line}.diff ; git add .osbs-logs/${line}.diff ; rhpkg commit -p -m "Updating dockerfile diff" ;;
-      esac
+      if [ "${FORCE}" == "TRUE" ] ; then
+        echo "  Force Option Selected - Assuming Continue"
+        mv -f .osbs-logs/${line}.diff.new .osbs-logs/${line}.diff ; git add .osbs-logs/${line}.diff ; rhpkg commit -p -m "Updating dockerfile diff" > /dev/null
+      else
+        echo "(c)ontinue [replace old diff], (i)gnore [leave old diff], (q)uit [exit script] : "
+        read choice < /dev/tty
+        case ${choice} in
+          c | C | continue ) mv -f .osbs-logs/${line}.diff.new .osbs-logs/${line}.diff ; git add .osbs-logs/${line}.diff ; rhpkg commit -p -m "Updating dockerfile diff" > /dev/null ;;
+          i | I | ignore ) rm -f .osbs-logs/${line}.diff.new ;;
+          q | Q | quit ) break;;
+          * ) echo "${choice} not and option.  Assuming ignore" ; rm -f .osbs-logs/${line}.diff.new ;;
+          #* ) echo "${choice} not and option.  Assuming continue" ;  mv -f .osbs-logs/${line}.diff.new .osbs-logs/${line}.diff ; git add .osbs-logs/${line}.diff ; rhpkg commit -p -m "Updating dockerfile diff" ;;
+        esac
+      fi
     fi
   done
+  echo "  ---- Checking Non-Dockerfile changes ----"
   find . -name ".git*" -prune -o -name ".osbs*" -prune -o -name "Dockerfile*" -prune -o -type f -print | while read line
   do
     diff -u ${line} ${workingdir}/${path}/${line}
   done
+  echo "  ---- Checking files added or removed ----"
   diff --brief -r ${workingdir}/${container} ${workingdir}/${path} | grep -v -e Dockerfile -e git -e osbs
   popd >/dev/null
 
@@ -241,13 +281,13 @@ while [[ "$#" -ge 1 ]]
 do
 key="$1"
 case $key in
-    everything_base | git_update_base | build_container_base)
+    everything | git_update | build_container)
       export list="${base_images_list}"
       export action="${key}"
       ;;
-    everything_s2i|git_update_s2i|build_container_s2i)
-      export list="${s2i_images_list}"
-      export action="${key}"
+    --group)
+      GROUP="$2"
+      shift
       ;;
     --version)
       version_version="$2"
@@ -268,47 +308,73 @@ case $key in
       DIST_GIT_BRANCH="$2"
       shift
       ;;
-    -v|--verbose)
-      export VERBOSE="TRUE"
-      ;;
+      -v|--verbose)
+        export VERBOSE="TRUE"
+        ;;
+      -f|--force)
+        export FORCE="TRUE"
+        ;;
     -h|--help)
       usage  # unknown option
       ;;
     *)
+      echo "Unknown Option: ${key}"
       usage  # unknown option
+      exit 4
       ;;
 esac
 shift # past argument or value
 done
 
+case $GROUP in
+  base)
+    export list="${base_images_list}"
+    ;;
+  s2i)
+    export list="${s2i_images_list}"
+    ;;
+  logging)
+    export list="${logging_images_list}"
+    ;;
+  metrics)
+    export list="${metrics_images_list}"
+    ;;
+  *)
+    echo "Unknown, or no group: ${GROUP}"
+    usage  # unknown group
+    exit 3
+    ;;
+esac
+
 workingdir=$(mktemp -d /var/tmp/rebuild-images-XXXXXX)
 pushd "${workingdir}" &>/dev/null
 mkdir -p logs/done
 echo "::None::" >> logs/finished
-touch logs/finished
+touch logs/buildfailed
 echo "Using working directory: ${workingdir}"
 
 echo "${list}" | while read spec
 do
   [ -z "$spec" ] && continue
+  export branch="${DIST_GIT_BRANCH}"
   export container=$(echo "$spec" | awk '{print $1}')
   export dependency=$(echo "$spec" | awk '{print $2}')
-  export branch=$(echo "$spec" | awk '{print $3}')
   export repo=$(echo "$spec" | awk '{print $4}')
   export path=$(echo "$spec" | awk '{print $5}')
   case "$action" in
-    build_container_base | build_container_s2i )
+    build_container )
       build_container
       ;;
-    git_update_base | git_update_s2i )
+    git_update )
        git_update
        ;;
-    everything_base | everything_s2i )
+    everything )
       git_update
       build_container
       ;;
     * )
       usage
+      exit 2
       ;;
   esac
 done
